requirements.txt
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
et_xmlfile==2.0.0
idna==3.11
msal==1.34.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.3
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
